requirements.txt
paho-mqtt>=1.6.1
requests>=2.31.0
boto3>=1.34.0
flask>=3.0.0
gunicorn>=21.0.0